feat(comment history): Implement feature to track comment edit history - ensure a comment has an edit history [Delivers #161278930]

+++ authors/apps/comments/views.py
@@ -20,6 +20,14 @@ from authors.apps.articles.permissions import IsAuthorOrReadOnly
 from authors.apps.articles.models import Article
 from authors.apps.likedislike.models import LikeDislike
 
+class CommentPagination(PageNumberPagination):
+    """
+    Article pagination support
+    """
+    page_size = 10
+    page_size_query_param = 'page_size'
+    max_page_size = 20
+
 
 class ArticleCommentAPIView(CreateAPIView):
     """A user can comment on an article"""
@@ -27,6 +35,7 @@ class ArticleCommentAPIView(CreateAPIView):
     permission_classes = (IsAuthenticatedOrReadOnly,)
     serializer_class = CommentSerializer
     renderer_classes = (CommentJSONRenderer,)
+    pagination_class = CommentPagination
     
 
     def post(self, request, slug=None):
@@ -47,16 +56,19 @@ class ArticleCommentAPIView(CreateAPIView):
         for comment in comment_set:
             serializer = CommentSerializer(comment)
             comments.append(serializer.data)
-            response = []
-            response.append({'comments': comments})
+            # response = []
+            # response.append({'comments': comments})
         commentsCount = len(comments)
         if commentsCount == 0:
             return Response({"Message":"There are no comments for this article"}, status=status.HTTP_200_OK) 
         elif commentsCount == 1:
-            return Response(response, status=status.HTTP_200_OK) 
+            return Response(comments, status=status.HTTP_200_OK) 
         else:
-            response.append({"commentsCount":commentsCount})
-            return Response(response, status=status.HTTP_200_OK) 
+            comments.append({"commentsCount":commentsCount})
+            page = self.paginate_queryset(comment_set)
+            if page is not None:
+                serializer = self.serializer_class(page, many=True)
+                return self.get_paginated_response(serializer.data)
 
 
 class ArticleCommentUpdateDeleteAPIView(RetrieveUpdateDestroyAPIView, CreateAPIView):
@@ -92,7 +104,8 @@ class ArticleCommentUpdateDeleteAPIView(RetrieveUpdateDestroyAPIView, CreateAPIV
     def update(self, request, slug=None, pk=None, **kwargs):
         """This function will update article"""
         self.permission_classes.append(IsAuthorOrReadOnly)
-        comment = self.get_object()
+        comment = get_object_or_404(Comment, pk=self.kwargs["id"])
+        self.check_object_permissions(self.request, comment)
         data = request.data
         serializer = self.serializer_class(comment, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
@@ -102,9 +115,11 @@ class ArticleCommentUpdateDeleteAPIView(RetrieveUpdateDestroyAPIView, CreateAPIV
     def delete(self, request, slug=None, pk=None, **kwargs):
         """This function will delete the article"""
         self.permission_classes.append(IsAuthorOrReadOnly)
-        comment = self.get_object()
+        comment = get_object_or_404(Comment, pk=self.kwargs["id"])
+        self.check_object_permissions(self.request, comment)
+        serializer = self.serializer_class(comment)
         comment.delete()
-        return Response({"message": {"Comment was deleted successfully"}}, status.HTTP_200_OK)
+        return Response("Successfully Deleted")
 
 class ImmediateCommentHistoryAPIView(RetrieveAPIView):
     """View to retrieve a previous comment"""
@@ -129,6 +144,8 @@ class ImmediateCommentHistoryAPIView(RetrieveAPIView):
 
 class AllCommentHistoryAPIView(RetrieveAPIView):
     """View to retrieve all previous comment"""
+    serializer_class = CommentSerializer
+    pagination_class = CommentPagination
 
     def get(self, request, slug=None, id=None):
         renderer_classes = (CommentJSONRenderer,)
@@ -140,8 +157,7 @@ class AllCommentHistoryAPIView(RetrieveAPIView):
             if not newcomments:
                 return Response({'Message':'This comment does not exist'}, status=status.HTTP_404_NOT_FOUND)
             for history_comment in newcomments:
-                serializer = CommentSerializer(history_comment)
-                comments.append(serializer.data)
-                response = []
-                response.append({'comments': comments})
-            return Response(response, status=status.HTTP_200_OK)
+                page = self.paginate_queryset(newcomments)
+                if page is not None:
+                    serializer = self.serializer_class(page, many=True)
+                    return self.get_paginated_response(serializer.data)
